+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.71.1
 
+- Fixes a regression introduced in v0.68.1 where incoming PreKey messages would fail to decrypt if the local device had archived the corresponding session.